+++ llm_generate_job/models/llm_generate_job.py
@@ -1,13 +1,12 @@
 import json
 import logging
-from datetime import datetime, timedelta
+from datetime import timedelta
 
 from odoo import api, fields, models
-from odoo.exceptions import UserError
-
 from odoo.addons.llm_mail_message_subtypes.const import (
     LLM_ASSISTANT_SUBTYPE_XMLID,
 )
+from odoo.exceptions import UserError
 
 _logger = logging.getLogger(__name__)
 
@@ -24,9 +23,9 @@ class LLMGenerateJob(models.Model):
 
     # Provider and model information
     provider_id = fields.Many2one(
-        "llm.provider", 
-        string="LLM Provider", 
-        required=True, 
+        "llm.provider",
+        string="LLM Provider",
+        required=True,
         tracking=True
     )
     model_id = fields.Many2one(
@@ -53,7 +52,7 @@ class LLMGenerateJob(models.Model):
         required=True,
         help="JSON data containing the generation parameters"
     )
-    
+
     webhook_url = fields.Char(
         string="Webhook URL",
         help="URL where the provider will send completion notification"
@@ -96,7 +95,7 @@ class LLMGenerateJob(models.Model):
         tracking=True,
         help="ID of the job on the provider's system (e.g., request_id from FAL)"
     )
-    
+
     gateway_request_id = fields.Char(
         string="Gateway Request ID",
         tracking=True,
@@ -125,7 +124,7 @@ class LLMGenerateJob(models.Model):
         string="Result Payload",
         help="JSON response from the provider containing the generated content"
     )
-    
+
     error_message = fields.Text(
         string="Error Message",
         tracking=True,
@@ -145,7 +144,7 @@ class LLMGenerateJob(models.Model):
         default=0,
         help="Number of times this job has been retried"
     )
-    
+
     max_retries = fields.Integer(
         string="Max Retries",
         default=3,
@@ -164,7 +163,7 @@ class LLMGenerateJob(models.Model):
         self.ensure_one()
         if self.state != 'draft':
             raise UserError("Only draft jobs can be submitted")
-        
+
         self._submit_to_provider()
         return True
 
@@ -173,14 +172,14 @@ class LLMGenerateJob(models.Model):
         self.ensure_one()
         if self.state in ['completed', 'failed', 'cancelled']:
             raise UserError("Cannot cancel a job that is already finished")
-        
+
         # Try to cancel with provider if it has an external ID
         if self.external_job_id and self.provider_id:
             try:
                 self.provider_id.cancel_generation_job(self.external_job_id)
             except Exception as e:
                 _logger.warning(f"Failed to cancel job {self.external_job_id} with provider: {e}")
-        
+
         self.write({
             'state': 'cancelled',
             'completed_date': fields.Datetime.now()
@@ -192,10 +191,10 @@ class LLMGenerateJob(models.Model):
         self.ensure_one()
         if self.state != 'failed':
             raise UserError("Only failed jobs can be retried")
-        
+
         if self.retry_count >= self.max_retries:
             raise UserError(f"Maximum number of retries ({self.max_retries}) exceeded")
-        
+
         self.write({
             'state': 'draft',
             'retry_count': self.retry_count + 1,
@@ -207,7 +206,7 @@ class LLMGenerateJob(models.Model):
             'completed_date': False,
             'result_payload': False,
         })
-        
+
         return self.action_submit()
 
     def action_hide_from_tree(self):
@@ -227,15 +226,15 @@ class LLMGenerateJob(models.Model):
     def _submit_to_provider(self):
         """Submit the job to the provider"""
         self.ensure_one()
-        
+
         # Generate webhook URL if not provided
         if not self.webhook_url:
             base_url = self.env['ir.config_parameter'].sudo().get_param('web.base.url')
             self.webhook_url = f"{base_url}/llm/generate_job/webhook/{self.id}"
-        
+
         # Submit to provider
         result = self.provider_id.submit_generation_job(self)
-        
+
         # Update job with provider response
         vals = {
             'state': 'submitted',
@@ -243,34 +242,41 @@ class LLMGenerateJob(models.Model):
             'external_job_id': result.get('request_id'),
             'gateway_request_id': result.get('gateway_request_id'),
         }
-        
+
         # If provider indicates immediate processing
         if result.get('status') == 'processing':
             vals['state'] = 'processing'
             vals['started_date'] = fields.Datetime.now()
-        
-        self.write(vals)
-        
-        _logger.info(f"Generation job '{self.name}' submitted successfully. "
-                    f"External job ID: {result.get('request_id')}")
 
-    def process_webhook_result(self, webhook_data):
+        self.write(vals)
+
+        _logger.info(f"Generation job '{self.name}' submitted successfully. "
+                     f"External job ID: {result.get('request_id')}")
+
+    def process_webhook_result(self, request, webhook_data):
         """Process webhook result from provider"""
         self.ensure_one()
-        
+        if not self.thread_id:
+            _logger.error(f"No thread found for job {self.id}")
+            return
+
+        if self.provider_id.is_webhook(request):
+            if not self.provider_id.verify_webhook(request, webhook_data):
+                _logger.error(f"Webhook verification failed for job {self.id}")
+                return {"status": "error", "message": "Verification failed"}
+
         status = webhook_data.get('status')
-        
+
         if status == 'OK':
-            # Success
             self.write({
                 'state': 'completed',
                 'completed_date': fields.Datetime.now(),
                 'result_payload': webhook_data.get('payload'),
             })
-            
+
             # Send result to thread
             self._send_result_to_thread(webhook_data.get('payload'))
-            
+
         elif status == 'ERROR':
             # Error
             self.write({
@@ -279,24 +285,24 @@ class LLMGenerateJob(models.Model):
                 'error_message': webhook_data.get('error', 'Unknown error'),
                 'result_payload': webhook_data.get('payload'),  # May contain error details
             })
-            
+
             # Notify thread about error
             self._send_error_to_thread(webhook_data.get('error', 'Unknown error'))
-        
+
         else:
             _logger.warning(f"Unknown webhook status '{status}' for job {self.id}")
 
     def _send_result_to_thread(self, payload):
         """Send successful generation result to the thread"""
         self.ensure_one()
-        
+
         if not self.thread_id:
             _logger.error(f"No thread found for job {self.id}")
             return
-        
-        try:            # Format the result for the thread
+
+        try:  # Format the result for the thread
             formatted_result = self._format_result_for_thread(payload)
-            
+
             # Post message to thread
             message = self.thread_id._post_message(
                 body=formatted_result.get('body', 'Generation completed'),
@@ -304,17 +310,17 @@ class LLMGenerateJob(models.Model):
                 subtype_xmlid=LLM_ASSISTANT_SUBTYPE_XMLID
             )
             self._send_realtime_notification(message)
-            
+
         except Exception as e:
             _logger.error(f"Failed to send result to thread {self.thread_id.id}: {e}")
 
     def _send_error_to_thread(self, error_message):
         """Send error notification to the thread"""
         self.ensure_one()
-        
+
         if not self.thread_id:
             return
-        
+
         try:
             body = f"❌ Generation failed: {error_message}"
             message = self.thread_id.message_post(
@@ -330,13 +336,13 @@ class LLMGenerateJob(models.Model):
     def _format_result_for_thread(self, payload):
         """Format the generation result for posting to thread"""
         self.ensure_one()
-        
+
         # This will need to be customized based on the type of generation
         # For now, handle common cases like images
-        
+
         body_parts = []
         attachment_ids = []
-        
+
         if isinstance(payload, dict):
             # Handle image generation results
             if 'images' in payload:
@@ -346,16 +352,16 @@ class LLMGenerateJob(models.Model):
                         # Create attachment from URL
                         try:
                             attachment = self._create_attachment_from_url(
-                                image['url'], 
-                                image.get('file_name', f'generated_image_{i+1}.png')
+                                image['url'],
+                                image.get('file_name', f'generated_image_{i + 1}.png')
                             )
                             if attachment:
                                 attachment_ids.append(attachment.id)
-                                body_parts.append(f"![Generated Image {i+1}]({image['url']})")
+                                body_parts.append(f"![Generated Image {i + 1}]({image['url']})")
                         except Exception as e:
                             _logger.error(f"Failed to create attachment from URL {image['url']}: {e}")
-                            body_parts.append(f"[Generated Image {i+1}]({image['url']})")
-            
+                            body_parts.append(f"[Generated Image {i + 1}]({image['url']})")
+
             # Handle video generation results  
             elif 'video' in payload:
                 body_parts.append("🎬 **Video generation completed!**")
@@ -372,16 +378,16 @@ class LLMGenerateJob(models.Model):
                     except Exception as e:
                         _logger.error(f"Failed to create attachment from URL {video['url']}: {e}")
                         body_parts.append(f"[Generated Video]({video['url']})")
-            
+
             # Handle generic results
             else:
                 body_parts.append("✅ **Generation completed!**")
                 body_parts.append(f"```json\n{json.dumps(payload, indent=2)}\n```")
-        
+
         else:
             body_parts.append("✅ **Generation completed!**")
             body_parts.append(str(payload))
-        
+
         return {
             'body': '\n\n'.join(body_parts),
             'attachment_ids': attachment_ids
@@ -405,17 +411,17 @@ class LLMGenerateJob(models.Model):
     def cleanup_completed_jobs(self, days_old=7):
         """Cleanup old completed jobs (called by cron)"""
         cutoff_date = fields.Datetime.now() - timedelta(days=days_old)
-        
+
         jobs_to_hide = self.search([
             ('state', 'in', ['completed', 'failed', 'cancelled']),
             ('completed_date', '<', cutoff_date),
             ('visible_in_tree', '=', True)
         ])
-        
+
         jobs_to_hide.write({'visible_in_tree': False})
-        
+
         _logger.info(f"Hidden {len(jobs_to_hide)} old generation jobs from tree view")
-        
+
         return len(jobs_to_hide)
 
     def _auto_hide_completed_jobs(self):
@@ -427,24 +433,24 @@ class LLMGenerateJob(models.Model):
     def write(self, vals):
         """Override write to auto-hide completed jobs"""
         result = super().write(vals)
-        
+
         # Auto-hide jobs when they complete
         if 'state' in vals and vals['state'] in ['completed', 'failed']:
             self._auto_hide_completed_jobs()
-        
+
         return result
 
     def _send_realtime_notification(self, message):
         """Send real-time notification to update the thread interface"""
         self.ensure_one()
-        
+
         if not message or not self.thread_id:
             return
-        
+
         try:
             # Send notification through Odoo's bus system
             channel = f"llm_thread_{self.thread_id.id}"
-            
+
             # Prepare notification payload
             notification = {
                 'type': 'new_message',
@@ -459,15 +465,15 @@ class LLMGenerateJob(models.Model):
                     'attachment_ids': message.attachment_ids.ids,
                 }
             }
-            
+
             # Send via bus using the Odoo 16 method
             self.env['bus.bus']._sendone(
-                [self.env.user.partner_id], 
-                'llm_thread_update', 
+                [self.env.user.partner_id],
+                'llm_thread_update',
                 notification
             )
-            
+
             _logger.info(f"Sent real-time notification for job {self.id} to channel {channel}")
-            
+
         except Exception as e:
             _logger.error(f"Failed to send real-time notification for job {self.id}: {e}")
